consolidated and renamed examples
