+++ examples/nrf24l01_ack_payload_test.py
@@ -1,0 +1,125 @@
+'''
+    Templated example of using the library to transmit
+    and retrieve custom automatic acknowledgment payloads.
+
+    Master transmits a dummy payload every second and prints the ACK payload.
+    Slave prints the received value and sends a dummy ACK payload.
+'''
+#  we'll be using pipe 0 to receive and transmit ACK packets
+
+import time
+import board
+import digitalio as dio
+from circuitpython_nrf24l01 import RF24
+
+# change these (digital output) pins accordingly
+ce = dio.DigitalInOut(board.D9)  # AKA board.CE1 on the rasberry pi
+csn = dio.DigitalInOut(board.D5)
+
+# using board.SPI() automatically selects the MCU's
+# available SPI pins, board.SCK, board.MOSI, board.MISO
+spi = board.SPI()  # init spi bus object
+
+# we'll be using the dynamic payload size feature (enabled by default)
+# the custom ACK payload feature is disabled by default
+# the custom ACK payload feature should not be enabled
+# during instantiation due to its singular use nature
+# meaning 1 ACK payload per 1 RX'd payload
+nrf = RF24(spi, csn, ce)
+
+# NOTE the the custom ACK payload feature will be enabled
+# automatically when you call load_ack() passing:
+# a buffer protocol object (bytearray) of
+# length ranging [1,32]. And pipe number always needs
+# to be an int ranging [0,5]
+
+# to enable the custom ACK payload feature
+nrf.ack = True  # False disables again
+
+# addresses needs to be in a buffer protocol object (bytearray)
+address = b'1Node'
+
+# payloads need to be in a buffer protocol object (bytearray)
+tx = b'Hello '
+
+# NOTE ACK payloads (like regular payloads and addresses)
+# need to be in a buffer protocol object (bytearray)
+ACK = b'World '
+
+
+def master(count=5):  # count = 5 will only transmit 5 packets
+    # recommended behavior is to keep in TX mode while idle
+    nrf.listen = False  # put radio in TX mode
+
+    # set address of RX node into a TX pipe
+    nrf.open_tx_pipe(address)
+
+    counter = count
+    while counter:
+        buffer = tx + bytes([counter + 48])  # output buffer
+        print("Sending (raw): {}".format(repr(buffer)))
+        # to read the ACK payload during TX mode we
+        # pass the parameter read_ack as True.
+        nrf.ack = True  # enable feature before send()
+        now = time.monotonic() * 1000  # start timer
+        result = nrf.send(buffer)  # becomes the response buffer
+        if result is None:
+            print('send() timed out')
+        elif not result:
+            print('send() failed')
+        else:
+            # print the received ACK that was automatically
+            # fetched and saved to "buffer" via send()
+            print('raw ACK: {}'.format(repr(result)))
+            # the ACK payload should now be in buffer
+        # print timer results despite transmission success
+        print('Transmission took',
+              time.monotonic() * 1000 - now, 'ms')
+        time.sleep(1)
+        counter -= 1
+
+# running slave to only fetch/receive count number of packets
+# count = 3 will mimic a full RX FIFO behavior via nrf.listen = False
+
+
+def slave(count=3):
+    # set address of TX node into an RX pipe. NOTE you MUST specify
+    # which pipe number to use for RX, we'll be using pipe 0
+    nrf.open_rx_pipe(0, address)
+
+    # put radio into RX mode, power it up, and set the first
+    # transmission's ACK payload and pipe number
+    nrf.listen = True
+    buffer = ACK + bytes([count + 48])
+    # we must set the ACK payload data and corresponding
+    # pipe number [0,5]
+    nrf.load_ack(buffer, 0)  # load ACK for first response
+
+    counter = count
+    start = time.monotonic()
+    while counter and (time.monotonic() - start) < (count * 2):
+        if nrf.any():
+            # this will listen indefinitely till counter == 0
+            counter -= 1
+            # print details about the received packet (if any)
+            print("Found {} bytes on pipe {}\
+                ".format(repr(nrf.any()), nrf.pipe()))
+            # retreive the received packet's payload
+            rx = nrf.recv()  # clears flags & empties RX FIFO
+            print("Received (raw): {}".format(repr(rx)))
+            start = time.monotonic()
+            if counter:  # Going again?
+                # build new ACK
+                buffer = ACK + bytes([counter + 48])
+                # load ACK for next response
+                nrf.load_ack(buffer, 0)
+
+    # recommended behavior is to keep in TX mode while idle
+    nrf.listen = False  # put radio in TX mode
+    nrf.flush_tx()  # flush any ACK payload
+
+
+print("""\
+    nRF24L01 ACK test\n\
+    Run slave() on receiver\n\
+    Run master() on transmitter""")
